[refactor] 라우터 페이지 구조 수정
