--- src/Home.js
@@ -1,15 +1,0 @@
-import React from 'react';
-import GoogleButton from './components/auth/GoogleButton';
-
-
-const Home = () => {
-    return (
-        <div>
-            <h1>Home 페이지 입니다</h1>
-            <GoogleButton />
-
-        </div>
-    );
-};
-
-export default Home;
