--- src/Investment.js
@@ -1,11 +1,0 @@
-import React from 'react';
-
-const Investment = () => {
-    return (
-        <div>
-            <h1>Investment 페이지 입니다</h1>
-        </div>
-    );
-};
-
-export default Investment;
