--- src/Ranking.js
@@ -1,11 +1,0 @@
-import React from 'react';
-
-const Ranking = () => {
-    return (
-        <div>
-            <h1>Ranking 페이지 입니다</h1>
-        </div>
-    );
-};
-
-export default Ranking;
