+++ src/components/home/index.js
@@ -1,0 +1,15 @@
+import React from 'react';
+import GoogleButton from '../auth/GoogleButton';
+
+
+const index = () => {
+    return (
+        <div>
+            <h1>Home 페이지 입니다</h1>
+            <GoogleButton />
+
+        </div>
+    );
+};
+
+export default index;
