+++ src/components/ranking/index.js
@@ -1,0 +1,11 @@
+import React from 'react';
+
+const index = () => {
+    return (
+        <div>
+            <h1>Ranking 페이지 입니다</h1>
+        </div>
+    );
+};
+
+export default index;
